+++ ereuse_devicehub/inventory/forms.py
@@ -86,12 +86,18 @@ DEVICES = {
         "Smartphone",
         "Cellphone",
     ],
+    "Drives & Storage": [
+        "All DataStorage",
+        "HardDrives",
+        "SolidStageDrive",
+    ],
 }
 
 COMPUTERS = ['Desktop', 'Laptop', 'Server', 'Computer']
 
 MONITORS = ["ComputerMonitor", "Monitor", "TelevisionSet", "Projector"]
 MOBILE = ["Mobile", "Tablet", "Smartphone", "Cellphone"]
+STORAGE = ["HardDrive", "SolidStateDrive"]
 
 
 class AdvancedSearchForm(FlaskForm):
@@ -175,8 +181,15 @@ class FilterForm(FlaskForm):
         elif "All Mobile" == self.device_type:
             filter_type = MOBILE
 
+        elif "All DataStorage" == self.device_type:
+            filter_type = STORAGE
+
         if filter_type:
             self.devices = self.devices.filter(Device.type.in_(filter_type))
+
+        # if self.device_type in STORAGE + ["All DataStorage"]:
+        # import pdb; pdb.set_trace()
+        # self.devices = self.devices.filter(Component.parent_id.is_(None))
 
         return self.devices.order_by(Device.updated.desc())
 
@@ -322,7 +335,7 @@ class NewDeviceForm(FlaskForm):
         default=1,
     )
     id_device_supplier = StringField('Id Supplier', [validators.Optional()])
-    phid = StringField('Placeholder Hardware identity (Phid)', [validators.Optional()])
+    id_device_internal = StringField('Id Internal', [validators.Optional()])
     pallet = StringField('Identity of pallet', [validators.Optional()])
     components = TextAreaField('Components', [validators.Optional()])
     info = TextAreaField('Info', [validators.Optional()])
@@ -382,7 +395,7 @@ class NewDeviceForm(FlaskForm):
         self.type.data = self._obj.type
         self.amount.render_kw = disabled
         self.id_device_supplier.data = self._obj.placeholder.id_device_supplier
-        self.phid.data = self._obj.placeholder.phid
+        self.id_device_internal.data = self._obj.placeholder.id_device_internal
         self.pallet.data = self._obj.placeholder.pallet
         self.info.data = self._obj.placeholder.info
         self.components.data = self._obj.placeholder.components
@@ -411,7 +424,7 @@ class NewDeviceForm(FlaskForm):
         if self._obj.placeholder.is_abstract:
             self.type.render_kw = disabled
             self.amount.render_kw = disabled
-            self.id_device_supplier.render_kw = disabled
+            # self.id_device_supplier.render_kw = disabled
             self.pallet.render_kw = disabled
             self.info.render_kw = disabled
             self.components.render_kw = disabled
@@ -474,28 +487,6 @@ class NewDeviceForm(FlaskForm):
                 int(meid, 16)
             except ValueError:
                 self.meid.errors = error
-                is_valid = False
-
-        if self.phid.data and self.amount.data == 1 and not self._obj:
-            dev = Placeholder.query.filter(
-                Placeholder.phid == self.phid.data, Placeholder.owner == g.user
-            ).first()
-            if dev:
-                msg = "Error, exist one Placeholder device with this PHID"
-                self.phid.errors = [msg]
-                is_valid = False
-
-        if (
-            self.phid.data
-            and self._obj
-            and self.phid.data != self._obj.placeholder.phid
-        ):
-            dev = Placeholder.query.filter(
-                Placeholder.phid == self.phid.data, Device.owner == g.user
-            ).first()
-            if dev:
-                msg = "Error, exist one Placeholder device with this PHID"
-                self.phid.errors = [msg]
                 is_valid = False
 
         if not is_valid:
@@ -579,8 +570,8 @@ class NewDeviceForm(FlaskForm):
 
     def reset_ids(self):
         if self.amount.data > 1:
-            self.phid.data = None
             self.id_device_supplier.data = None
+            self.id_device_internal.data = None
             self.serial_number.data = None
             self.part_number.data = None
             self.sku.data = None
@@ -590,8 +581,8 @@ class NewDeviceForm(FlaskForm):
     def get_placeholder(self):
         self.placeholder = Placeholder(
             **{
-                'phid': self.phid.data or None,
                 'id_device_supplier': self.id_device_supplier.data,
+                'id_device_internal': self.id_device_internal.data,
                 'info': self.info.data,
                 'components': self.components.data,
                 'pallet': self.pallet.data,
@@ -601,10 +592,12 @@ class NewDeviceForm(FlaskForm):
         return self.placeholder
 
     def edit_device(self):
-        self._obj.placeholder.phid = self.phid.data or self._obj.placeholder.phid
         if not self._obj.placeholder.is_abstract:
             self._obj.placeholder.id_device_supplier = (
                 self.id_device_supplier.data or None
+            )
+            self._obj.placeholder.id_device_internal = (
+                self.id_device_internal.data or None
             )
             self._obj.placeholder.info = self.info.data or None
             self._obj.placeholder.components = self.components.data or None
@@ -1279,6 +1272,12 @@ class TradeDocumentForm(FlaskForm):
 
 
 class TransferForm(FlaskForm):
+    lot_name = StringField(
+        'Lot Name',
+        [validators.DataRequired()],
+        render_kw={'class': "form-control"},
+        description="You need put a lot name",
+    )
     code = StringField(
         'Code',
         [validators.DataRequired()],
@@ -1295,14 +1294,14 @@ class TransferForm(FlaskForm):
     def __init__(self, *args, **kwargs):
         self._type = kwargs.get('type')
         lot_id = kwargs.pop('lot_id', None)
-        self._tmp_lot = Lot.query.filter(Lot.id == lot_id).one()
+        self._tmp_lot = None
+        if lot_id:
+            self._tmp_lot = Lot.query.filter(Lot.id == lot_id).one()
         super().__init__(*args, **kwargs)
         self._obj = None
 
     def validate(self, extra_validators=None):
         is_valid = super().validate(extra_validators)
-        if not self._tmp_lot:
-            return False
 
         if self._type and self.type.data not in ['incoming', 'outgoing']:
             return False
@@ -1323,8 +1322,10 @@ class TransferForm(FlaskForm):
         return self._obj
 
     def set_obj(self):
-        self.newlot = Lot(name=self._tmp_lot.name)
-        self.newlot.devices = self._tmp_lot.devices
+        name = self.lot_name.data
+        self.newlot = Lot(name=name)
+        if self._tmp_lot:
+            self.newlot.devices = self._tmp_lot.devices
         db.session.add(self.newlot)
 
         self._obj = Transfer(lot=self.newlot)
@@ -1355,6 +1356,7 @@ class EditTransferForm(TransferForm):
             self.code.data = self._obj.code
             self.description.data = self._obj.description
             self.date.data = self._obj.date
+            self.lot_name.data = self._obj.lot.name
 
     def validate(self, extra_validators=None):
         is_valid = super().validate(extra_validators)
@@ -1366,6 +1368,7 @@ class EditTransferForm(TransferForm):
 
     def set_obj(self, commit=True):
         self.populate_obj(self._obj)
+        self._obj.lot.name = self.lot_name.data
 
 
 class NotesForm(FlaskForm):
@@ -1514,9 +1517,7 @@ class UploadPlaceholderForm(FlaskForm):
             else:
                 self.source = "Excel File: {}".format(_file.filename)
             try:
-                data = (
-                    pd.read_excel(_file, converters={'Phid': str}).fillna('').to_dict()
-                )
+                data = pd.read_excel(_file).fillna('').to_dict()
             except ValueError:
                 txt = ["File don't have a correct format"]
                 self.placeholder_file.errors = txt
@@ -1538,12 +1539,12 @@ class UploadPlaceholderForm(FlaskForm):
             return False
 
         header = [
-            'Phid',
             'Model',
             'Manufacturer',
             'Serial Number',
             'Part Number',
             'Id device Supplier',
+            'Id device Internal',
             'Pallet',
             'Info',
         ]
@@ -1557,32 +1558,7 @@ class UploadPlaceholderForm(FlaskForm):
         self.placeholders = []
         schema = SnapshotSchema()
         self.path_snapshots = {}
-        for i in data['Phid'].keys():
-            placeholder = None
-            data['Phid'][i] = str(data['Phid'][i])
-            if data['Phid'][i]:
-                placeholder = Placeholder.query.filter_by(phid=data['Phid'][i]).first()
-
-            # update one
-            if placeholder:
-                self.dev_update += 1
-                device = placeholder.device
-                device.model = "{}".format(data['Model'][i]).lower()
-                device.manufacturer = "{}".format(data['Manufacturer'][i]).lower()
-                device.serial_number = "{}".format(data['Serial Number'][i]).lower()
-                device.part_number = "{}".format(data['Part Number'][i]).lower()
-                placeholder.id_device_supplier = "{}".format(
-                    data['Id device Supplier'][i]
-                )
-                placeholder.pallet = "{}".format(data['Pallet'][i])
-                placeholder.info = "{}".format(data['Info'][i])
-
-                placeholder_log = PlaceholdersLog(
-                    type="Update", source=self.source, placeholder=device.placeholder
-                )
-                self.placeholders.append((device, placeholder_log))
-                continue
-
+        for i in data['Model'].keys():
             # create a new one
             json_snapshot = {
                 'type': 'Snapshot',
@@ -1597,8 +1573,8 @@ class UploadPlaceholderForm(FlaskForm):
                 },
             }
             json_placeholder = {
-                'phid': data['Phid'][i] or None,
                 'id_device_supplier': data['Id device Supplier'][i],
+                'id_device_internal': data['Id device Internal'][i],
                 'pallet': data['Pallet'][i],
                 'info': data['Info'][i],
                 'is_abstract': False,
@@ -1635,7 +1611,6 @@ class EditPlaceholderForm(FlaskForm):
     serial_number = StringField('Serial Number', [validators.Optional()])
     part_number = StringField('Part Number', [validators.Optional()])
     id_device_supplier = StringField('Id Supplier', [validators.Optional()])
-    phid = StringField('Phid', [validators.DataRequired()])
     pallet = StringField('Pallet', [validators.Optional()])
     info = StringField('Info', [validators.Optional()])