+++ ereuse_devicehub/workbench/views.py
@@ -3,33 +3,44 @@ import time
 import flask
 from flask import Blueprint
 from flask import current_app as app
-from flask import g, make_response, request
+from flask import g, make_response, request, url_for
+from flask.views import View
 from flask_login import login_required
 
 from ereuse_devicehub import auth
 from ereuse_devicehub.db import db
+from ereuse_devicehub.resources.device.models import Placeholder
 from ereuse_devicehub.resources.enums import SessionType
 from ereuse_devicehub.resources.user.models import Session
 from ereuse_devicehub.views import GenericMixin
 from ereuse_devicehub.workbench import isos
+from ereuse_devicehub.workbench.forms import KangarooForm
 
 workbench = Blueprint('workbench', __name__, url_prefix='/workbench')
 
 
 class SettingsView(GenericMixin):
     decorators = [login_required]
+    methods = ['GET', 'POST']
     template_name = 'workbench/settings.html'
-    page_title = "Workbench"
+    page_title = "Snapshots"
 
     def dispatch_request(self):
         self.get_context()
+        form_kangaroo = KangarooForm()
         self.context.update(
             {
                 'page_title': self.page_title,
                 'demo': g.user.email == app.config['EMAIL_DEMO'],
-                'iso': isos,
+                'iso_demo': isos['demo'],
+                'iso_register': isos['register'],
+                'iso_erease': isos['erease'],
+                'form': form_kangaroo,
             }
         )
+
+        if form_kangaroo.validate_on_submit():
+            form_kangaroo.save()
 
         self.opt = request.values.get('opt')
         if self.opt in ['register', 'erease_basic', 'erease_sectors']:
@@ -86,4 +97,24 @@ class SettingsView(GenericMixin):
         return token
 
 
+class ErasureHostView(View):
+    decorators = [login_required]
+    methods = ['GET']
+
+    def dispatch_request(self, id):
+        self.placeholder = (
+            Placeholder.query.filter(Placeholder.id == id)
+            .filter(Placeholder.kangaroo.is_(True))
+            .filter(Placeholder.owner_id == g.user.id)
+            .one()
+        )
+        self.placeholder.kangaroo = False
+        db.session.commit()
+
+        return flask.redirect(url_for('workbench.settings'))
+
+
 workbench.add_url_rule('/', view_func=SettingsView.as_view('settings'))
+workbench.add_url_rule(
+    '/erasure_host/<int:id>/', view_func=ErasureHostView.as_view('erasure_host')
+)